Cambiando tipo entero del código C de int64_t a int.

+++ packages/pytiger2c/types/integertype.py
@@ -17,4 +17,4 @@ class IntegerType(BasicType):
         Inicializa la clase representando el tipo array.
         """
         super(IntegerType, self).__init__()
-        self._code_type = 'int64_t'
+        self._code_type = 'int'
